\name{kaps-package}
\alias{kaps-package}
\docType{package}
\title{
	Functions for K-adaptive partitioning and recursive partitioning for censored survival data.
}
\description{
  This package provides some routines for K-adaptive partitioning and recursive partitioning algorithms for censored survival data.
}
\details{
\tabular{ll}{
Package: \tab kaps\cr
Type: \tab Package\cr
Version: \tab 1.0.0\cr
Date: \tab 2012-03-15\cr
License: \tab GPL version 2 or newer\cr
LazyLoad: \tab no\cr
}
This package contains some routines for K-adaptive partitioning and recursive partitioning algorithms when the data is censored, especially the application to SEER (Surveillance Epidemiology and End Results).
}
\author{
Soo-Heang Eo <hanansh@korea.ac.kr> and HyungJun Cho <hj4cho@korea.ac.kr>
}

\references{
Eo, SH, Hong, S-M and Cho, H (2012+). K-Adaptive Partitioning for Survival data with the application to SEER, in progression.  \cr


}
\seealso{
    \code{\link{kaps}}, \code{\link{lrtree}}
}
\keyword{ package }