\name{km.curve}
\alias{km.curve}
\title{Plot Kaplan-Meire survival curves}
\description{
    Plot the Kaplan-Meire survival curves for terminal nodes or subgroups on a grapic devide.
}
\usage{
	km.curve(object, 
	x.lab = c(0,24,48,72,96,120, 144, 168, 192, 216, 240), ...)
}

\arguments{
  \item{object}{an object corresponding to "apss" or "Tree" S4 formulation classes. }
  \item{x.lab}{ X labels specified as arguments}
  \item{\dots}{ other arguments for plot object. See \link[graphics:plot]{plot} for details.}
}

\details{
  This function provides Kaplan-Meire survival curves for terminal nodes or subgroups generated by \code{\link{lrtree}}. 
}
\seealso{
	\code{\link{lrtree}}
}
\keyword{methods}
