\name{labels}
\docType{methods}
\alias{labels}
\alias{labels,Tree-method}
\title{Methods for Function \code{labels} for \code{lrtree}}
\description{Create Split Labels for an "Tree".}
\usage{
	\S4method{labels}{Tree}(object, type = "diagram")
}
\arguments{
  \item{object}{object of S4 class \code{Tree}.}
  \item{type}{types of labels to project in tree diagram and results.}
}
\keyword{methods}
