\name{show}
\docType{methods}
\alias{show}
\alias{show,apss-method}
\alias{show,Tree-method}
\title{Methods for Function \code{show} for the "apss" and "Tree" classes. }
\usage{
\S4method{show}{apss}(object)
\S4method{show}{Tree}(object)
}
\description{show simple summary for \code{\link{kaps}} and \code{\link{lrtree}} objects.  It provide just the call and the structure of test-statistic.
}
\arguments{
  \item{object}{object of S4 class \code{Tree} or \code{apss}.}
}
\seealso{
  \code{\link{lrtree}} and \code{\linkS4class{Tree}} class \cr
  \code{\link{kaps}} and \code{\linkS4class{apss}} class
}
\keyword{methods}
