\name{toy}
\alias{toy}
\title{toy example}
\description{
    This is an toy example similar to SEER dataset.
}
\arguments{
  \item{meta}{the number of metastatic lymph nodes}
  \item{exam}{the number of examined lymph nodes}
  \item{status}{censoring status}
  \item{time}{ time to event}
}

\keyword{datasets}
