\name{print}
\docType{methods}
\alias{print,kaps-method}
\title{Print an object "kaps" with specific information about K}
\usage{
\S4method{print}{kaps}(x, K)
}
\description{
It functions like \code{show} but the only difference is the output with the specific information about K.}
\arguments{
  \item{x}{an object from \code{kaps}}
  \item{K}{a scalar object to determine the number of subgroups K. If missing, the estimated subgroup K is selected.}
}
\seealso{
  \code{\link{kaps}}
}
\keyword{methods}
