\name{show}
\docType{methods}
\alias{show}
\alias{show,kaps-method}
\title{Methods for Function \code{show} for the "kaps" S4 class. }
\usage{
\S4method{show}{kaps}(object)
}
\description{show simple summary for \code{\link{kaps}} object.  It provide just the call and the structure of test-statistic.
}
\arguments{
  \item{object}{object of S4 class \code{kaps}.}
}
\seealso{
  \code{\link{kaps}} and \code{\linkS4class{kaps}} class
}
\keyword{methods}
